REFACTOR: Create HomeModule and move pages
