+++ frontend/src/app/app-routing.module.ts
@@ -1,17 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { CreatePartnerComponent } from './pages/create-partner/create-partner.component';
-import { CreateUserComponent } from './pages/create-user/create-user.component';
-import { HomeComponent } from './pages/home/home.component';
-import { NotFoundComponent } from './shared/components/not-found/not-found.component';
 
-const routes: Routes = [
-  { path: 'home', component: HomeComponent },
-  { path: 'create-user', component: CreateUserComponent },
-  { path: 'create-partner', component: CreatePartnerComponent },
-  { path: '', redirectTo: '/home', pathMatch: 'full' },
-  { path: '**', component: NotFoundComponent },
-];
+const routes: Routes = [];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
