+++ frontend/src/app/app.module.ts
@@ -4,23 +4,16 @@ import { registerLocaleData } from '@angular/common';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 
-import { HomeComponent } from './pages/home/home.component';
-import { CreateUserComponent } from './pages/create-user/create-user.component';
-import { CreatePartnerComponent } from './pages/create-partner/create-partner.component';
-
 import { SharedModule } from './shared/shared.module';
 
 import ptBr from '@angular/common/locales/pt';
+import { HomeModule } from './home/home.module';
+
 registerLocaleData(ptBr);
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    HomeComponent,
-    CreateUserComponent,
-    CreatePartnerComponent,
-  ],
-  imports: [AppRoutingModule, SharedModule],
+  declarations: [AppComponent],
+  imports: [AppRoutingModule, SharedModule, HomeModule],
   providers: [{ provide: LOCALE_ID, useValue: 'pt' }],
   bootstrap: [AppComponent],
 })
