+++ frontend/src/app/home/home.component.scss
