+++ frontend/src/app/home/home.component.ts
@@ -2,7 +2,7 @@ import { UserServiceService } from 'src/app/shared/services/user-service.service
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { Table } from 'primeng/table';
 import { PrimeNGConfig } from 'primeng/api';
-import { ShowRoutesComponent } from '../../shared/components/show-routes/show-routes.component';
+import { ShowRoutesComponent } from '../shared/components/show-routes/show-routes.component';
 
 @Component({
   selector: 'app-home',
