+++ frontend/src/app/home/pages/create-partner/create-partner.component.scss
